productSold method

+++ Backend-MarketPlace/app/Http/Controllers/ProductController.php
@@ -24,37 +24,35 @@ class ProductController extends Controller
         $start_date = $request->start_date;
         $end_date = $request->end_date;
         $state = $request->state;
-        $user_id = auth()->user()->id;
-        
-        /**
-         * $access_token = $request->access_token;
-         * if (userProfile() == null) {
-         *    throw new RuntimeException("You have to be connected to add a product");
-         * }
-         */
-
-        return 
-            Product::create(['label' => $label, 'description' => $description, 'initial_price' => $initial_price, 'current_price' => $current_price, 'start_date' => $start_date, 'end_date' => $end_date, 'state' => $state, 'user_id' => $user_id]); 
+        $owner_id = auth()->user()->id;
+        $image = $request->image;
+        return Product::create(['label' => $label, 'description' => $description, 'initial_price' => $initial_price, 'current_price' => $current_price, 'start_date' => $start_date, 'end_date' => $end_date, 'state' => $state, 'owner_id' => $owner_id, 'client_id' => $owner_id, 'image' => $image]); 
     }
 
     public function editProduct(Request $request)
     {
-        $user_id = $request->user_id;
+        $user_id = Product::find($request->id)->owner_id;
         $client_id = auth()->user()->id;
         if ($user_id!= $client_id) {
             return response()->json([
-                'error' => 'You are not allowed to edit this product'
+                'error' => 'You are not allowed to edit product of other user'
             ], 401);
         }else{
             return Product::find($request->id)->update($request->all());
         }
-        
     }
-    
+
     public function bidProduct(Request $request)
     {
-        $user_id = $request->user_id;
+        $product = Product::find($request->id);
+        $user_id = $product->owner_id;
         $client_id = auth()->user()->id;
+
+        if ($product->state == 'sold'){
+            return response()->json([
+                'error' => 'Product is sold'
+            ], 401);
+        }
         if ($user_id == $client_id) {
             throw new RuntimeException('You are not allowed to bid your own product');
         }
@@ -72,7 +70,25 @@ class ProductController extends Controller
 
     public function deleteProduct(Request $request)
     {
+        $user_id = Product::find($request->id)->owner_id;
+        $client_id = auth()->user()->id;
+        if ($user_id != $client_id) {
+            throw new RuntimeException('You are not allowed to delete products of other users');
+        }
+
         $id = $request->id;
         return Product::find($id)->delete();
+    }
+
+    public function productSold(Request $request)
+    {
+        $product = Product::find($request->id);
+
+        if ($product->state == 'sold') {
+            throw new RuntimeException('The product is already sold');
+        }
+        $state = 'sold';
+        $product->update(['state' => $state]);
+        return $product;
     }
 }
